EK-108: changed names of fieldComponents to "namex" for transitition to PascalCase
