+++ gatsby_fp/src/components/fieldComponents/generalSectionx.js
@@ -1,0 +1,12 @@
+import React from "react"
+
+export default function GeneralSection({ fieldLabel, fieldData }) {
+  return (
+    <>
+      <div className="sectiontitle">{fieldLabel}</div>
+      <div>
+        <p dangerouslySetInnerHTML={{ __html: fieldData }}></p>
+      </div>
+    </>
+  )
+}
