--- gatsby_fp/src/components/fieldComponents/profSocieties.js
@@ -1,77 +1,0 @@
-import React from "react"
-const shortid = require("shortid")
-
-export default function ProfSocieties({ profSocietiesList }) {
-  return (
-    <>
-      <div className="sectiontitle">Professional Societies</div>
-      <div>
-        <ul>
-          {profSocietiesList.map(element => (
-            <li
-              key={shortid.generate()}
-              dangerouslySetInnerHTML={{
-                __html: ` 
-              ${
-                element.org
-                  ? `
-                org: ${element.org}<br> `
-                  : ``
-              } 
-              ${
-                element.orgabbr
-                  ? `
-                orgabbr: ${element.orgabbr}<br> `
-                  : ``
-              } 
-              ${
-                element.status
-                  ? `
-                orgabbr: ${element.orgabbr}<br> `
-                  : ``
-              } 
-              ${
-                element.dtm_start
-                  ? `
-                dtm_start: ${element.dtm_start}<br> `
-                  : ``
-              } 
-              ${
-                element.dtd_start
-                  ? `
-                dtd_start: ${element.dtd_start}<br> `
-                  : ``
-              } 
-              ${
-                element.dty_start
-                  ? `
-                dty_start: ${element.dty_start}<br> `
-                  : ``
-              }
-              ${
-                element.dtm_end
-                  ? `
-                dtm_end: ${element.dtm_end}<br> `
-                  : ``
-              } 
-              ${
-                element.dtd_end
-                  ? `
-                dtd_end: ${element.dtd_end}<br> `
-                  : ``
-              } 
-              ${
-                element.dty_end
-                  ? `
-                dty_end: ${element.dty_end}<br> `
-                  : ``
-              }
-          `,
-              }}
-            ></li>
-          ))}
-        </ul>
-      </div>
-    </>
-  )
-}
